refactor: fixture 위치 수정
