+++ src/test/java/com/woowacamp/soolsool/acceptance/fixture/RestLiquorFixture.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.fixture;
+package com.woowacamp.soolsool.acceptance.fixture;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
@@ -11,10 +11,10 @@ public abstract class RestLiquorFixture {
 
     public static Long 술_등록_새로_판매중(String accessToken) {
         LiquorSaveRequest request = new LiquorSaveRequest(
-                "SOJU", "GYEONGGI_DO", "ON_SALE",
-                "새로", "3000", "롯데", "/soju-url",
-                100, 12.0, 300,
-                LocalDateTime.now().plusYears(5L)
+            "SOJU", "GYEONGGI_DO", "ON_SALE",
+            "새로", "3000", "롯데", "/soju-url",
+            100, 12.0, 300,
+            LocalDateTime.now().plusYears(5L)
         );
 
         return addLiquor(accessToken, request);
@@ -22,10 +22,10 @@ public abstract class RestLiquorFixture {
 
     public static Long 술_등록_ETC_경기도_하이트_진로_판매중지(String accessToken) {
         LiquorSaveRequest request = new LiquorSaveRequest(
-                "ETC", "GYEONGGI_DO", "STOPPED",
-                "하이트", "4000", "진로", "/beer-url",
-                200, 24.0, 600,
-                LocalDateTime.now().plusYears(5L)
+            "ETC", "GYEONGGI_DO", "STOPPED",
+            "하이트", "4000", "진로", "/beer-url",
+            200, 24.0, 600,
+            LocalDateTime.now().plusYears(5L)
         );
 
         return addLiquor(accessToken, request);
@@ -33,10 +33,10 @@ public abstract class RestLiquorFixture {
 
     public static Long 술_등록_과일주_전라북도_얼음딸기주_우영미_판매중(String accessToken) {
         LiquorSaveRequest request = new LiquorSaveRequest(
-                "BERRY", "JEOLLABUK_DO", "ON_SALE",
-                "얼음딸기주", "4500", "우영미", "/strawberry-url",
-                20, 14.0, 400,
-                LocalDateTime.now().plusYears(5L)
+            "BERRY", "JEOLLABUK_DO", "ON_SALE",
+            "얼음딸기주", "4500", "우영미", "/strawberry-url",
+            20, 14.0, 400,
+            LocalDateTime.now().plusYears(5L)
         );
 
         return addLiquor(accessToken, request);
@@ -44,13 +44,13 @@ public abstract class RestLiquorFixture {
 
     private static Long addLiquor(String accessToken, LiquorSaveRequest request) {
         return Long.valueOf(RestAssured
-                .given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer" + accessToken)
-                .contentType(APPLICATION_JSON_VALUE)
-                .accept(APPLICATION_JSON_VALUE)
-                .body(request)
-                .when().post("/liquors")
-                .then().log().all()
-                .extract().header("Location").split("/")[2]);
+            .given().log().all()
+            .header(HttpHeaders.AUTHORIZATION, "Bearer" + accessToken)
+            .contentType(APPLICATION_JSON_VALUE)
+            .accept(APPLICATION_JSON_VALUE)
+            .body(request)
+            .when().post("/liquors")
+            .then().log().all()
+            .extract().header("Location").split("/")[2]);
     }
 }
